Removing win32console from gemspec and adding Rakefile/Gemfile. Removing binary .rb extension, and setting it to be executable.
The Rakefile includes bundler tasks, run `rake -T` to see them. Make it
easier to install on your local machine and release.

The Gemfile allows any contributor to just `bundle install` and be set
with any dependencies. I linked the Gemfile to the gemspec so there's no
duplication.

I included the `curb` gem to the gemspec, and removed the win32console
since that's platform specific. We could either mention it in the README
or use a hook to check the platform at package time (which require an
extension to Rubygem.)

We could also benefit from an .rvmrc file.